+++ src/zandronum/src/f_wipe.cpp
@@ -287,7 +287,7 @@ bool wipe_doBurn (int ticks)
 		{
 			int fglevel;
 
-			fglevel = burnarray[(firex>>FRACBITS)+(firey>>FRACBITS)*FIREWIDTH] / 2;
+			fglevel = burnarray[(int)((firex>>FRACBITS)+(firey>>FRACBITS)*FIREWIDTH)] / 2;
 			if (fglevel >= 63)
 			{
 				to[x] = fromnew[x];
@@ -344,7 +344,7 @@ bool wipe_doFade (int ticks)
 		int x, y;
 		fixed_t bglevel = 64 - fade;
 		DWORD *fg2rgb = Col2RGB8[fade];
-		DWORD *bg2rgb = Col2RGB8[bglevel];
+		DWORD *bg2rgb = Col2RGB8[(int)(bglevel)];
 		BYTE *fromnew = (BYTE *)wipe_scr_end;
 		BYTE *fromold = (BYTE *)wipe_scr_start;
 		BYTE *to = screen->GetBuffer();
